qt-lib, qt-qml: Move FileFinder and QrcParser to qt-lib
This enables use of these classes from other extensions.
