+++ qt-lib/package.json
@@ -34,6 +34,7 @@
   "dependencies": {
     "@vscode/extension-telemetry": "^1.1.0",
     "async": "^3.2.6",
+    "fast-xml-parser": "^5.3.4",
     "module-alias": "^2.2.3",
     "typescript": "^5.6.3",
     "untildify": "^5.0.0",
