+++ qt-lib/src/qrc-parser.ts
