+++ qt-qml/package.json
@@ -455,7 +455,6 @@
     "@vscode/debugprotocol": "^1.68.0",
     "@vscode/l10n": "^0.0.16",
     "async-mutex": "^0.5.0",
-    "fast-xml-parser": "^5.0.8",
     "get-port": "^7.1.0",
     "module-alias": "^2.2.3",
     "promise-socket": "^8.0.0",
